requirements.txt
APScheduler==3.10.1
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
pycoingecko==3.1.0
pyTelegramBotAPI==4.10.0
python-dotenv==1.0.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
six==1.16.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
